+++ winery-operate/src/main/java/com/changfa/frame/website/controller/role/RoleController.java
@@ -125,7 +125,7 @@ public class RoleController {
             if(roleDTO.getId().equals(null) || roleDTO.getId() == null){
                 return JsonReturnUtil.getJsonIntReturn(1, "角色信息有误");
             }
-            Role roleOld = roleRepository.getOne(roleDTO.getId());
+            Role roleOld = roleRepository.getOne(roleDTO.getId().longValue());
             if (roleOld == null) {
                 return JsonReturnUtil.getJsonIntReturn(4, "找不到角色id" + roleDTO.getId());
             }
@@ -206,7 +206,7 @@ public class RoleController {
             if (adminUser == null) {
                 return JsonReturnUtil.getJsonIntReturn(1, "找不到token" + token);
             } else {
-                Role role = roleRepository.getOne(roleId);
+                Role role = roleRepository.getOne(roleId.longValue());
                 if(null == role){
                     return JsonReturnUtil.getJsonIntReturn(1, "角色信息有误" + token);
                 }
@@ -229,7 +229,7 @@ public class RoleController {
             if (adminUser == null) {
                 return JsonReturnUtil.getJsonIntReturn(1, "找不到token" + token);
             } else {
-                Role role = roleRepository.getOne(roleId);
+                Role role = roleRepository.getOne(roleId.longValue());
                 if(null == role){
                     return JsonReturnUtil.getJsonIntReturn(1, "角色信息有误" + token);
                 }
